+++ UmaMusumeAPI/Models/Tables/AnnounceCharacter.cs
@@ -6,7 +6,7 @@ namespace UmaMusumeAPI.Models.Tables
     public class AnnounceCharacter
     {
         public int Id { get; set; }
-        public long MovieId { get; set; }
-        public long CanEndSecond { get; set; }
+        public int MovieId { get; set; }
+        public int CanEndSecond { get; set; }
     }
 }
