+++ UmaMusumeAPI/Models/Tables/BackgroundData.cs
@@ -6,14 +6,14 @@ namespace UmaMusumeAPI.Models.Tables
     public class BackgroundData
     {
         public int Id { get; set; }
-        public long BgId { get; set; }
-        public long BgSub { get; set; }
+        public int BgId { get; set; }
+        public int BgSub { get; set; }
         public string CueName { get; set; }
         public string SheetName { get; set; }
         public string BusParamSetName { get; set; }
-        public long PosX { get; set; }
-        public long PosY { get; set; }
-        public long Width { get; set; }
-        public long Height { get; set; }
+        public int PosX { get; set; }
+        public int PosY { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
     }
 }
