+++ UmaMusumeAPI/Models/Tables/BackgroundOffsetData.cs
@@ -6,12 +6,12 @@ namespace UmaMusumeAPI.Models.Tables
     public class BackgroundOffsetData
     {
         public int Id { get; set; }
-        public long BgId { get; set; }
-        public long BgSub { get; set; }
-        public long OffsetId { get; set; }
-        public long PosX { get; set; }
-        public long PosY { get; set; }
-        public long Width { get; set; }
-        public long Height { get; set; }
+        public int BgId { get; set; }
+        public int BgSub { get; set; }
+        public int OffsetId { get; set; }
+        public int PosX { get; set; }
+        public int PosY { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
     }
 }
