+++ UmaMusumeAPI/Models/Tables/CampaignCharaStorySchedule.cs
@@ -2,8 +2,8 @@
 {
     public class CampaignCharaStorySchedule
     {
-        public long CampaignId { get; set; }
-        public long CharaId { get; set; }
-        public long StoryId { get; set; }
+        public int CampaignId { get; set; }
+        public int CharaId { get; set; }
+        public int StoryId { get; set; }
     }
 }
