+++ UmaMusumeAPI/Models/Tables/ChallengeMatchRawPoint.cs
@@ -6,12 +6,12 @@ namespace UmaMusumeAPI.Models.Tables
     public class ChallengeMatchRawPoint
     {
         public int Id { get; set; }
-        public long Priority { get; set; }
-        public long ConditionType { get; set; }
-        public long ConditionValue1 { get; set; }
-        public long ConditionValue2 { get; set; }
-        public long Point { get; set; }
-        public long RacePointNameId { get; set; }
-        public long SortOrder { get; set; }
+        public int Priority { get; set; }
+        public int ConditionType { get; set; }
+        public int ConditionValue1 { get; set; }
+        public int ConditionValue2 { get; set; }
+        public int Point { get; set; }
+        public int RacePointNameId { get; set; }
+        public int SortOrder { get; set; }
     }
 }
