+++ UmaMusumeAPI/Models/Tables/ChampionsNewsTrainerWininfo.cs
@@ -6,10 +6,10 @@ namespace UmaMusumeAPI.Models.Tables
     public class ChampionsNewsTrainerWininfo
     {
         public int Id { get; set; }
-        public long RoundId { get; set; }
-        public long RaceCountType { get; set; }
-        public long WinMin { get; set; }
-        public long WinMax { get; set; }
-        public long TrainerTextGroup { get; set; }
+        public int RoundId { get; set; }
+        public int RaceCountType { get; set; }
+        public int WinMin { get; set; }
+        public int WinMax { get; set; }
+        public int TrainerTextGroup { get; set; }
     }
 }
